Test `CustomType` quotation
